共同編集の仮実装
